Redirect if anon user accesses hymnal page, add show/hide password
This commit adds the ability for users to toggle between showing and hiding their password on the login page. It also includes a loading component that is displayed when a user attempts to access a book without being logged in. Additionally, it redirects users to the login page if they try to access a book without being logged in and then returns them back to the book after logging in.